+++ odoo/local-src/delivery_rounds_refill/models/report_stock_refill_arrange.py
@@ -50,6 +50,11 @@ class ReportStockRefillArrange(models.Model):
     planned_count = fields.Integer('Planned outgoing count')
     immediate_qty = fields.Integer('Immediate outgoing qty')
     immediate_count = fields.Integer('Immediate outgoing count')
+    pending_round_reserved_qty = fields.Integer(
+        'Reserved qty in bin',
+        help="Quantity in bin, reserved for delivery rounds which "
+        "are not started",
+    )
     average_qty = fields.Integer('Average outgoing qty')
     average_count = fields.Integer('Average outgoing count')